Skip assignment to flaghash associative array on Bash 3
Prior to this commit the attempts to assign a value to the flaghash
associative array would fail on OSX (Bash3) as associative arrays were
only introduced in Bash4.

This commit skips assigning values to the flaghash unless running on
Bash >3.

The flaghash handling was added in #205, with the workaround in 801364a
only partially fixing the problem.

Signed-off-by: John McCabe <john@johnmccabe.net>